#fix fixed up 3 sonarqube warnings

+++ identityprovider/src/main/java/nz/ac/canterbury/seng302/identityprovider/authentication/JwtTokenUtil.java
@@ -26,7 +26,7 @@ public class JwtTokenUtil implements Serializable {
 	private JwtTokenUtil() {}
 
 	private static final long serialVersionUID = -2550185165626007488L;
-	public static final long JWT_TOKEN_VALIDITY = 5 * 60 * 60;
+	public static final long JWT_TOKEN_VALIDITY = 5 * 60 * (long) 60;
 
 	public static final String ROLE_CLAIM_TYPE = "role";
 	public static final String NAME_CLAIM_TYPE = "name";
